moved gameClasses into models and created keep-alive cronjob
